added crowdranking oembed endpoint
This endpoint supports `xml`, `json` and `jsonp`. `js` or `json` with and `callback` parameter are also treated as `jsonp`.
Supported parameters are: `maxwidth`, `maxheight` and `lang`. There are other service specific parameters that I have to document some day.

It returns the type `rich` with `html` containing an iframe. It also returns an `url` that points to the iframe source and two attributes I "stole" from flickr: `web_page` (a link to the ranking on crowdranking.com) and `web_page_short_url` (short-ish; this url might get shorter in the future).